feat: add separate page just for the prompt, no data
