+++ src/routes/(app)/project/[id]/(page)/prompt/+page.svelte
@@ -1,0 +1,59 @@
+<script lang="ts">
+	import { updatePrompt } from '$lib/api';
+	import PromptOptions from '$lib/components/options/PromptOptions.svelte';
+	import PaywallPopup from '$lib/components/popups/PaywallPopup.svelte';
+	import PromptBar from '$lib/components/prompt/PromptBar.svelte';
+	import PromptEditor from '$lib/components/prompt/PromptEditor.svelte';
+	import type { Tables } from '$lib/supabase';
+	import { ChevronDown, ChevronUp } from 'lucide-svelte';
+
+	let { data } = $props();
+
+	let showPaywall = $state(false);
+	let prompt = $state(data.prompt);
+	let showOptions = $state(false);
+	let editedPrompt = $state({ ...prompt });
+	let suggestionApplied = $state(false);
+	let hoveredSuggestion: Tables<'suggestions'> | null = $state(null);
+
+	function setPrompt() {
+		updatePrompt(editedPrompt).then((r) => {
+			if (r === null) {
+				showPaywall = true;
+				return;
+			}
+			prompt = r;
+			showOptions = false;
+			suggestionApplied = false;
+			editedPrompt = { ...prompt };
+		});
+	}
+</script>
+
+{#if showPaywall}
+	<PaywallPopup
+		onclose={() => (showPaywall = false)}
+		message="You have reached your monthly limit of 100 prompts."
+	/>
+{/if}
+<PromptBar bind:prompt projectId={data.project.id || ''} userStatus={data.user.status} />
+<div class="flex w-full flex-col px-6 py-4">
+	<div class="mb-2 flex items-end justify-between">
+		<h1>Prompt</h1>
+		<button
+			class="flex items-center gap-1 opacity-40 transition-all hover:opacity-100"
+			onclick={() => (showOptions = !showOptions)}
+		>
+			<span class="text-black">Model Options</span>
+			{#if showOptions}
+				<ChevronUp class="h-5 w-5" />
+			{:else}
+				<ChevronDown class="h-5 w-5" />
+			{/if}
+		</button>
+	</div>
+	{#if showOptions}
+		<PromptOptions bind:prompt={editedPrompt} userStatus={data.user.status} />
+	{/if}
+	<PromptEditor {prompt} {hoveredSuggestion} {setPrompt} bind:suggestionApplied bind:editedPrompt />
+</div>
